[Linux] Add FlViewWidget, make FlView an interface
Implements https://github.com/flutter/flutter/issues/157906

This allows for custom embedders to implement the new FlView interface
and retain some compatibility with existing plugins.
